restructure viz section
