+++ lib/util.js
@@ -18,7 +18,7 @@ const debug = require('debug')('0x')
 const launch = require('opn')
 const { promisify } = require('util')
 const profLogConvert = require('./prof-log-convert')
-const html = require('./visualize/html')
+const html = require('../visualizer/html')
 
 const phases = [
   (frames) => frames,
@@ -125,8 +125,8 @@ function createBundle () {
       multistream([
         browserify({standalone: 'd3'}).add(require.resolve('d3')).bundle(),
         browserify({standalone: 'hsl'}).add(require.resolve('hsl-to-rgb-for-reals')).bundle(),
-        browserify({standalone: 'ui'}).add(require.resolve('d3-fg')).bundle(),
-        browserify({standalone: 'visualize'}).add(path.join(__dirname, 'visualize')).bundle()
+        browserify({standalone: 'fg'}).add(require.resolve('d3-fg')).bundle(),
+        browserify({standalone: 'visualizer'}).add(path.resolve(__dirname, '..', 'visualizer')).bundle()
       ]),
       concat(resolve),
       (err) => err && reject(err)
@@ -137,7 +137,7 @@ function createBundle () {
 async function render (args, {json, pid, folder}) {
   const { name, title } = args
   debug('converted stacks to intermediate format')
-  const script = `${await createBundle()}\nvisualize(${JSON.stringify(json)}, ${JSON.stringify({title})})`
+  const script = `${await createBundle()}\nvisualizer(${JSON.stringify(json)}, ${JSON.stringify({title})})`
 
   const htmlPath = determineHtmlPath(args, {pid, folder})
   const opts = {
