+++ visualizer/cmp/controls.js
